[editorial] Mismatched quotes and framing as blockquote (chapter 19)
This entire chapter is a document that the narrator reads. In the original, this is indicated by setting the entire chapter in quotes. This is further complicated by the fact that after the first paragraph, the document is mostly a transcript of court testimony. Within that testimony, the speaker also recounts dialogue.

The mixture of single and double quotes is treated inconsistently in the original source.

I elected to solve this by setting the entire chapter as a blockquote (semanticated as a letter—a decision I'm less certain about)