[Editorial] Capitalization
"this" does appear as lowercase in the original, perhaps because the phrase is a sentence fragment. The rest of the questions that follow begin with a capital, but then again they are all complete setences. I think it is less disruptive to have it capitalized. But not a hill I'm willing to die on (or even to fight for).