[Editorial] Standardize spelling of incontestably
Original source spells it with an <i>a</i> twice and once with an <i>i</i>. Transcriber changes one of the a's to an i but misses the other. Both versions seem to be have been used interchangeably at the time of original publication but the a version is now considered correct.